R6 Class to print components of CV from data

R6 Class to print components of CV from data

R6 Class to print components of CV from data

Details

This class is initiated at the head of your CV or Resume Rmarkdown file and then through various print_* methods, builds the various components.

Public fields

position_data

data frame of positions by row with columns:

section What type of position entry,

title Title of entry,

loc Where the position took place,

institution Institution the position was associated with,

start Start year of position

end End year of position,

in_resume Logical to filter what entries should be included in a resume (Not used for CV mode),

description_{1,2,...} Free form text fields to be added as description bullets. description_2 description_3

skills

data frame with two columns:

skill ID of skill

level Relative numeric level for skill

text_blocks

data frame with two columns:

loc Where this text lock is going in CV

text Actual text to be placed.

contact_info

data frame with three columns:

loc What the contact point is for (e.g. email)

icon Font-awesome 4 icon id for this contact point (e.g. "envelope")

contact Actual contact info such as nick@test.com.

pdf_mode

Is the output being rendered into a pdf? Aka do links need to be stripped?

html_location

Where will the html version of your CV be hosted?

pdf_location

Where will the pdf version of your CV be hosted?

Methods

Public methods

CV_Printer$new()

CV_Printer$set_pdf_mode()

CV_Printer$print_section()

CV_Printer$print_text_block()

CV_Printer$print_skill_bars()

CV_Printer$print_links()

CV_Printer$print_contact_info()

CV_Printer$print_link_to_other_format()

CV_Printer$set_style()

CV_Printer$clone()

Method new()

Create a CV_Printer object.

Usage

CV_Printer$new(
  data_location,
  pdf_mode = FALSE,
  html_location,
  pdf_location,
  position_entry_template = default_position_entry_template,
  sheet_is_publicly_readable = TRUE
)

Arguments

data_location

Path of the spreadsheets holding all your data. This can be either a URL to a google sheet with multiple sheets containing the four data types or a path to a folder containing four .csvs with the neccesary data.

pdf_mode

Is the output being rendered into a pdf? Aka do links need to be stripped?

html_location

What location will the HTML version of this CV be hosted at?

pdf_location

What location will the PDF of this CV be hosted at?

position_entry_template

A glue template for building position entries.

sheet_is_publicly_readable

If you're using google sheets for data, is the sheet publicly available? (Makes authorization easier.)

source_location

Where is the code to build your CV hosted?

Returns

A new CV_Printer object.

Method set_pdf_mode()

Turn on pdf mode for class. Useful for when the class is cached to avoid re-downloading data.

Usage

CV_Printer$set_pdf_mode(pdf_mode = TRUE)

Arguments

pdf_mode

Are we turning PDF mode on?

Method print_section()

Take a position data frame and the section id desired and prints the section to markdown.

Usage

CV_Printer$print_section(section_id)

Arguments

section_id

ID of the positions section to be printed as encoded by the section column of the positions table

Method print_text_block()

Prints out text block identified by a given label.

Usage

CV_Printer$print_text_block(label)

Arguments

label

ID of the text block to print as encoded in label column of text_blocks table.

Method print_skill_bars()

Construct a bar chart of skills

Usage

CV_Printer$print_skill_bars(out_of = 5)

Arguments

out_of

The relative maximum for skills. Used to set what a fully filled in skill bar is.

Method print_links()

List of all links in document labeled by their superscript integer.

Usage

CV_Printer$print_links()

Method print_contact_info()

Contact information section with icons

Usage

CV_Printer$print_contact_info()

Method print_link_to_other_format()

Small addendum that links to pdf version of CV if currently HTML and HTML if currently PDF.

Usage

CV_Printer$print_link_to_other_format()

Method set_style()

Appends some styles specific to PDF output. glue template for building position entries Internal array holding all the links that have been stripped in the order they were stripped.

Usage

CV_Printer$set_style()

Method clone()

The objects of this class are cloneable with this method.

Usage

CV_Printer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.