Use CSVs for storing data

Sets up examples of the four CSVs needed for building CV

use_csv_data_storage(folder_name = "data", create_output_dir = TRUE)

Arguments

folder_name

Name of the folder you want csvs stored in relative to current working directory

create_output_dir

If the requested output directory is missing should it be created?

Value

A new folder <folder_name>/ with entries.csv, text_blocks.csv, language_skills.csv, and contact_info.csv in it. working directory.

Examples

# Make a temp directory for placing files
# This would be a real location for a typical situation
temp_dir <- fs::dir_create(fs::path(tempdir(), "cv_w_csvs"))

datadrivencv::use_csv_data_storage(
  folder_name = fs::path(temp_dir, "csv_data"),
  create_output_dir = TRUE
)
#> [1] "Copied CSVs to  /var/folders/rp/ttzsjwxs6bx0x__xbb402xv80000gn/T/RtmpgXuDwI/cv_w_csvs/csv_data"

list.files(fs::path(temp_dir, "csv_data"))
#> [1] "contact_info.csv"    "entries.csv"         "language_skills.csv"
#> [4] "text_blocks.csv"