Use Data Driven CV template

Sets up the .Rmd file for a data-driven cv in current working directory. Also adds css file for current CV so style can be custommized.

use_datadriven_cv(
  full_name = "Sarah Arcos",
  data_location = system.file("sample_data/", package = "datadrivencv"),
  pdf_location = "https://github.com/nstrayer/cv/raw/master/strayer_cv.pdf",
  html_location = "nickstrayer.me/datadrivencv/",
  source_location = "https://github.com/nstrayer/datadrivencv",
  which_files = "all",
  output_dir = getwd(),
  create_output_dir = FALSE,
  use_network_logo = TRUE,
  open_files = TRUE
)

Arguments

full_name

Your full name, used in title of document and header

data_location

Path of the spreadsheets holding all your data. This can be either a URL to a google sheet with multiple sheets containing the four data types or a path to a folder containing four .csvs with the neccesary data. See use_csv_data_storage() for help setting up these .csvs.

pdf_location

What location will the PDF of this CV be hosted at?

html_location

What location will the HTML version of this CV be hosted at?

source_location

Where is the code to build your CV hosted?

which_files

What files should be placed? Takes a vector of possible values c("cv.rmd", "dd_cv.css", "render_cv.r", "cv_printing_functions.r") or "all" for everything. This can be used to incrementally update the printing functions or CSS without loosing customizations you've made to other files.

output_dir

Where should the files be placed? Defaults to your current working directory

create_output_dir

If the requested output directory is missing should it be created?

use_network_logo

Should logo be an interactive network based on your CV data? Note that this uses the function build_network_logo() so will introduce a dependency on this package.

open_files

Should the added files be opened after creation?

Value

cv.rmd, dd_cv.css, render_cv.r, and cv_printing_functions.r

written to the current working directory.

Examples

# Make a temp directory for placing files
# This would be a real location for a typical situation
temp_dir <- fs::dir_create(fs::path(tempdir(), "my_cv"))

use_datadriven_cv(
  full_name = "Nick Strayer",
  data_location = "https://docs.google.com/spreadsheets/d/14MQICF2F8-vf8CKPF1m4lyGKO6_thG-4aSwat1e2TWc",
  pdf_location = "https://github.com/nstrayer/cv/raw/master/strayer_cv.pdf",
  html_location = "nickstrayer.me/cv/",
  source_location = "https://github.com/nstrayer/cv",
  output_dir = temp_dir,
  open_files = FALSE
)

# Files should be where they were requested
list.files(temp_dir)
#> [1] "cv.rmd"                  "cv_printing_functions.r"
#> [3] "dd_cv.css"               "dd_resume.css"          
#> [5] "render_cv.r"             "resume.rmd"